+++ diesel/src/expression_methods/mod.rs
@@ -8,7 +8,7 @@ mod bool_expression_methods;
 mod eq_all;
 mod escape_expression_methods;
 mod global_expression_methods;
-#[cfg(any(feature = "sqlite", feature = "postgres_backend"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "postgres_backend"))]
 pub(crate) mod json_expression_methods;
 mod text_expression_methods;
 
@@ -23,7 +23,7 @@ pub use self::global_expression_methods::{
     ExpressionMethods, NullableExpressionMethods, UntypedExpressionMethods,
 };
 #[doc(inline)]
-#[cfg(any(feature = "sqlite", feature = "postgres_backend"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "postgres_backend"))]
 pub use self::json_expression_methods::{AnyJsonExpressionMethods, JsonIndex};
 #[doc(inline)]
 pub use self::text_expression_methods::TextExpressionMethods;
@@ -45,6 +45,6 @@ pub use crate::expression::functions::aggregate_expressions::frame_clause::{
 #[doc(inline)]
 pub use crate::pg::expression::expression_methods::*;
 
-#[cfg(feature = "sqlite")]
+#[cfg(feature = "__sqlite-shared")]
 #[doc(inline)]
 pub use crate::sqlite::expression::expression_methods::*;