+++ diesel/src/query_builder/select_statement/boxed.rs
@@ -1,4 +1,4 @@
-use std::marker::PhantomData;
+use core::marker::PhantomData;
 
 use crate::backend::{DieselReserveSpecialization, sql_dialect};
 use crate::dsl::AsExprOf;
@@ -603,7 +603,7 @@ mod tests {
         #[cfg(feature = "postgres")]
         assert_boxed_query_send!(crate::pg::Pg);
 
-        #[cfg(feature = "sqlite")]
+        #[cfg(feature = "__sqlite-shared")]
         assert_boxed_query_send!(crate::sqlite::Sqlite);
 
         #[cfg(feature = "mysql")]