+++ diesel/src/query_dsl/save_changes_dsl.rs
@@ -1,19 +1,19 @@
-#[cfg(any(feature = "sqlite", feature = "postgres", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "postgres", feature = "mysql"))]
 use crate::Table;
 use crate::associations::HasTable;
-#[cfg(any(feature = "sqlite", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "mysql"))]
 use crate::associations::Identifiable;
 use crate::connection::Connection;
-#[cfg(any(feature = "sqlite", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "mysql"))]
 use crate::dsl::Find;
-#[cfg(any(feature = "sqlite", feature = "postgres", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "postgres", feature = "mysql"))]
 use crate::dsl::Update;
-#[cfg(any(feature = "sqlite", feature = "postgres", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "postgres", feature = "mysql"))]
 use crate::expression::{MixedAggregates, ValidGrouping, is_aggregate};
 use crate::query_builder::{AsChangeset, IntoUpdateTarget};
-#[cfg(any(feature = "sqlite", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "mysql"))]
 use crate::query_dsl::methods::{ExecuteDsl, FindDsl};
-#[cfg(any(feature = "sqlite", feature = "postgres", feature = "mysql"))]
+#[cfg(any(feature = "__sqlite-shared", feature = "postgres", feature = "mysql"))]
 use crate::query_dsl::{LoadQuery, RunQueryDsl};
 use crate::result::QueryResult;
 
@@ -49,10 +49,10 @@ where
     }
 }
 
-#[cfg(feature = "sqlite")]
+#[cfg(feature = "__sqlite-shared")]
 use crate::sqlite::SqliteConnection;
 
-#[cfg(feature = "sqlite")]
+#[cfg(feature = "__sqlite-shared")]
 impl<'b, Changes, Output> UpdateAndFetchResults<Changes, Output> for SqliteConnection
 where
     Changes: Copy + Identifiable,